Added icons, and stuffs for the Netherlands translation.

+++ tools/createqm.sh
@@ -7,11 +7,23 @@
 #
 ##############################################
 
-if [ x"$TSLANG" == "x" ]; then
-    TSLANG=ja_JP
+PACKAGELISTFILE="tools/packagelist.txt"
+
+if [ -z $POOTLE_LANG ] 
+then
+    echo "The POOTLE_LANG variable is not set!"
+    echo "Please set it with export POOTLE_LANG=hu before calling this script!"
+    echo "The variable should match with the language code on the pootle."
+    exit -1
 fi
 
-PACKAGELISTFILE="tools/packagelist.txt"
+if [ -z $QM_SUFFIX ] 
+then
+    echo "The QM_SUFFIX variable is not set!"
+    echo "Please set it with export QM_SUFFIX=hu before calling this script!"
+    echo "This variable should be set to the preferred qm file suffix on the target device"
+    exit -1
+fi
 
 rm -rf usr/share/translations/*
 i=0
@@ -19,7 +31,5 @@ TS_COUNT=$(wc -l < $PACKAGELISTFILE)
 while IFS='' read -r line || [[ -n "$line" ]]; do
     FILENAME=$(echo "$line" | sed "s/.*\///")
     FILENAME=$(echo $FILENAME | awk -F"." '{print $1}')
-    lrelease -removeidentical -idbased translations/$line -qm usr/share/translations/$FILENAME-${TSLANG}.qm
+    lrelease -idbased translations/$POOTLE_LANG/$line -qm usr/share/translations/$FILENAME-${QM_SUFFIX}.qm
 done < $PACKAGELISTFILE
-
-find usr/share/translations -name "*-${TSLANG}.qm"
